Make the computed collection properties on ListParameters truly read only.
This makes it so the ASP.NET Core model binder doesn't see them and think that it can incorrectly add items into the collections. Caching was also added to OrderByList as it was previously not present, causing the dictionary to be reconstructed at least 3 times per /list request.

This lack of true read-onlyness was allowing requests with extra parameters that shouldn't have been getting read (FieldList) to end up actually modifying the cached field list collection.

For example:
http://localhost:5000/api/Person/list?Page=&PageSize=&OrderBy=&OrderByDescending=&Fields=name&FieldList=System.Collections.Generic.List`1[System.String]&OrderByList=System.Collections.Generic.Dictionary`2[System.String%2cIntelliTect.Coalesce.SortDirection]&Search=&Includes=&DataSource=

These extra parameters might end up here if someone is directly serializing a ListParameters object to make an outgoing request from a C# application.