Flip even more bits so that coalesce-vue works with both `node12` module resolution and "classic" resolution.
In particular, point "main" back at the ESM build (this is the way it was essentially forever up until 15e2240aacba41c4e27e661a77876bdcce3c8d08). This restores the old behavior, while the presence of the `exports` field enables the new dual cjs/esm behavior for newer TS/NodeJS versions that know to look at this field.

The field "module" was not ever doing anything. I don't think this is a real, respected field.