feat!: don't treat abstract entities as if they have no DbSet.
This adds partial support for #8 - TPH/TPT scenarios - by allowing abstract base entities to have a read-only API exposed. This allows these entities to be chosen in dropdowns, but doesn't add much beyond that. The frontend is still entirely unaware of the hierarchy and has no knowledge of the discriminator.